Update README.md
DISCLAIMER
This extension was written many years ago as fun and now its not work (or may be not work as it has to be). Sorry, if you want to use but read my comment first: https://github.com/recoders/chessbot/issues/72#issuecomment-631733282

Anyway, I know that GitHub is full of bright talents who wants to make it work, so - go ahead. I can support you, but only by words, not by new code.

So if you want to help - create fork and do commit. Any other issues will be closed as irrelevant.

Thank you and take care of yourself and your family! v